Imports: only process covers from supported hosts in `load()`
Because of the HTTP Proxy requirements, only known cover hosts are used
when importing via `load()` (e.g. from `/isbn` or via `/api/import`).

`'cover'` values not found in `ALLOWED_COVER_HOSTS` are simply ignored
during import to prevent timeouts when trying to connect to hosts not
added to the proxy.