5831/feature/unified read button dropdown (#9544)
## Implement Unified Read Button Changes

Creates a Unified Read Button Dropper that works across the Books Page, Authors Page, and Search Page. This includes the Book Page editions table.

Currently uses a conditional check to avoid rendering in carousels because (a) carousels have not yet been made #10472 Edition-Aware and also because the styling is broken.

The dropper shows the right primary action based on the case (e.g. Borrow, Read, Special Access)

The dropdown options (which display in cases other than carousels) show `locate` and `listen` where applicable, with icons, and will soon show `search inside` and `download` options.

We switched the Locate feature to use a new /borrow?action=locate method if an OCLC or ISBN identifier exists for an edition and simplified the code logic away from WorldcatURL.html.

---------

Co-authored-by: Mek <michael.karpeles@gmail.com>