Create separate package.json for storybook stories
I found storybook was taking a decent junk of time when npm installing,
so split it out into a separate package.json . I set the npm commands to
auto-install the new package.json though, so there should be no breaking
changes in anyone's flow. The `npm run storybook` will work as before.

Only potentialy breaking change is that when storybook is built, it'll
build in stories/storybook-static instead of the root-level
storybook-static/.