Speed up `npm ci` ~55% by creating separate package.json for storybook stories (#10612)
* Create separate package.json for storybook stories

I found storybook was taking a decent junk of time when npm installing,
so split it out into a separate package.json . I set the npm commands to
auto-install the new package.json though, so there should be no breaking
changes in anyone's flow. The `npm run storybook` will work as before.

Only potentialy breaking change is that when storybook is built, it'll
build in stories/storybook-static instead of the root-level
storybook-static/.

* Fix stories/ subdir not finding .babelrc

* Disable package-lock.json for stories/ subdir

Since these dependencies are only ever used for local development, they
are incredibly unlikely to need the rigor associated with
package-lock.json, which is usually more important for production-facing
code.