Use our Wikidata IDs for reading stat demographics
Previously, we’ve done a reverse lookup of Open Library IDs in Wikidata
to fetch demographic data for the reading stats page.
This is prone to fail as OLIDs “change” due to merges and not getting
updated in Wikidata.

This commit changes the logic to instead use the Wikidata IDs we have on
our side and use those to look up data from Wikidata, regardless of what
OLIDs are assigned there—if any.

This change makes it more immediately possible for OL librarians to
fix authors missing Wikidata links (or being wrongly linked), without
having to (additionally) learn how to edit Wikidata.

Fixes https://github.com/internetarchive/openlibrary/issues/6552