Include OCLC ids on the Add Book page (#10006)
* Include OCLC ids in the Add Book page’s dropdown

OCLC ids will usually be available for titles that do not have
an ISBN to identify them and are also used a lot in Wikipedia
to reference publications.

This uses the same ordering used in the `$popular` array for
adding identifiers on the book editing page.

Fixes https://github.com/internetarchive/openlibrary/issues/8719

* Add validation of OCLC ids on Add Book page

Based mostly on the validation of LCCN identifiers, this also
generalises the `displayLccnError()` function into a generic
`displayIdentifierError()` since there wasn’t anything LCCN specific
about it and I didn’t want to WET the code by copy/pasting the code for
a `displayOclcError()` function. :)