Initialize modals using `dialog.js` module
Adds `dialog--open` class and `aria-controls` to the
dialog opener affordances (the "Edit" and "Other" links).
These changes cause the dialog openers to be initialized
in the main `index.js` flow, in the typical manner.

`open-modal` custom events and the `openModal` method
were removed from `CheckInComponents.js, as the dialog
openers are now initialized in `index.js``.

Adds `initDialogClosers` function to `dialog.js` for
adding listeners to elements that were created
client-side.