Reduce scope of script
Modifies the script to only write a preference
to the store if the following conditions are
true:

1. The preference object contains a `pda` key
2. The preference object is not already written to the store

The `_fetch_preference_keys` function returns
all of the keys that meet these criteria.
These keys are then passed to `copy_preferences_to_store` in
batches of 1,000.

`copy_preferences_to_store` now returns a list of keys that
were, for any reason, not written to the store.  These keys
are reinserted into the list of keys to be written to store
for retry in a later `copy_preferences_to_store` call.