Tweak solr parameters to reduce heap usage
Scott noticed that the fields `title_sort` and `name` were appearing
frequently in Solr's fieldCache and consuming a lot of space (~1.8G for
`title`). He noted that the documentation recommends for things like
sorting/faceting, docValues are recommended -- which solr.TextField
doesn't support. But _does_ support them is the aptly names
solr.SortableTextField!

So we switched to that for our `title_sort` field, and for the author
`name` field we switched to instead sort by `name_str` which should
behave the same.

Co-authored-by: Scott Barnes <scottreidbarnes@gmail.com>