Fix: make /import/batch/approve work
Closes #11227

This commit ensures the `batch_id` value passed into the query is the
same as the value of `batch_id` as it is passed in from the endpoint.

Previously, this value was always `1`. E.g., visiting
`/import/batch/approve/3` would show:
```
batch_id = '3'
0.0 (1):
            UPDATE import_item
            SET status = 'pending'
            WHERE batch_id = 1 AND status = 'needs_review';
```

This change in this commit was tested for SQL injection by setting `batch_id` to '3 OR true':
```
psycopg2.errors.InvalidTextRepresentation: invalid input syntax for integer: "3 OR true"
LINE 4:             WHERE batch_id = '3 OR true' AND status = 'needs...
                                     ^
```